% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cos_}
\alias{cos_}
\title{Fast parallel version of cosinus (when vector size >= 50000)}
\usage{
cos_(x, paralen = 50000L)
}
\arguments{
\item{x}{vector a numeric values}

\item{paralen}{the minimum length of x to use parallel computation (50000
by default)}
}
\value{
a numeric vector with the cos(x) values
}
\description{
Fast parallel version of cosinus (when vector size >= 50000)
}
