% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_}
\alias{log_}
\title{Fast parallel version of log(x, base) (when vector size >= 50000)}
\usage{
log_(x, base = 2.718282, paralen = 50000L)
}
\arguments{
\item{x}{vector a numeric values}

\item{base}{the base of the logarithm (e = exp(1) by default)}

\item{paralen}{the minimum length of x to use parallel computation (50000
by default)}
}
\value{
a numeric vector with the log_base(x) values
}
\description{
Fast parallel version of log(x, base) (when vector size >= 50000)
}
