% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Round_}
\alias{Round_}
\alias{ceiling_}
\alias{floor_}
\alias{trunc_}
\title{Rounding of Numbers (Fast Parallel Version)}
\usage{
ceiling_(x, paralen = 50000L)

floor_(x, paralen = 50000L)

trunc_(x, paralen = 50000L)
}
\arguments{
\item{x}{vector of numeric values}

\item{paralen}{the minimum length of \code{x} to use parallel computation (50000
by default)}
}
\value{
A numeric vector, matrix, or data frame with the transformed values.
}
\description{
Fast version of rounding (when vector size >= 50000).
\code{\link[=ceiling_]{ceiling_()}} takes a single numeric argument x and returns a numeric vector
containing the smallest integers not less than the corresponding elements of
\code{x}. It is similar to \code{\link[=ceiling]{ceiling()}}.

\code{\link[=floor_]{floor_()}} takes a single numeric argument \code{x} and returns a numeric vector
containing the largest integers not greater than the corresponding elements of
\code{x}. It is similar to \code{\link[=floor]{floor()}}.

\code{\link[=trunc_]{trunc_()}} takes a single numeric argument \code{x} and returns a numeric vector
containing the integers formed by truncating the values in \code{x} toward 0. It
is similar to \code{\link[=trunc]{trunc()}}.
}
\details{
They are \strong{not} generic functions.
In case of factor, Date, POSIXt, difftime, complex, or S4 objects, the
function delegates to the base function. For data frames, no dispatching is
done and direct column-wise computation is performed. All attributes are
preserved.
}
\examples{
floor_(c(1.23, 4.56, -7.89))
}
\seealso{
\code{\link[=ceiling]{ceiling()}}, \code{\link[=floor]{floor()}}, \code{\link[=trunc]{trunc()}}
}
