% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svFast-package.R
\docType{package}
\name{svFast-package}
\alias{svFast}
\alias{svFast-package}
\title{'SciViews' - Fast and Parallelized Versions of Various R Functions}
\description{
A series of fast math and stat functions, parallelized using RcppParallel.
}
\details{
This section should provide a more detailed overview of how to use the
package, including the most important functions.
}
\section{Math functions}{

More info about parallelized math functions
}

\section{Stat functions}{

More info about stat functions...
}

\references{
This optional section can contain literature or other references for
background information.
}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org} (\href{https://orcid.org/0000-0002-2694-9471}{ORCID})

}
\keyword{internal}
