% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Trig_}
\alias{Trig_}
\alias{cos_}
\alias{sin_}
\alias{tan_}
\title{Trigonometric Functions (Fast Parallel Version)}
\usage{
cos_(x, paralen = 50000L)

sin_(x, paralen = 50000L)

tan_(x, paralen = 50000L)
}
\arguments{
\item{x}{vector of numeric values}

\item{paralen}{the minimum length of \code{x} to use parallel computation (50000
by default)}
}
\value{
A numeric vector, matrix, or data frame with the transformed values.
}
\description{
Fast version of trigonometric functions (when vector size >= 50000). They
respectively compute the cosine, sine, tangent, arc-cosine, arc-sine,
arc-tangent, and the two-argument arc-tangent.

\code{cospi_(x)}, \code{sinpi_(x)}, and \code{tanpi_(x)}, compute \code{cos(pi*x)},
\code{sin(pi*x)}, and \code{tan(pi*x)}.
}
\details{
They are \strong{not} generic functions.
In case of factor, Date, POSIXt, difftime, complex, or S4 objects, the
function delegates to the base function. For data frames, no dispatching is
done and direct column-wise computation is performed. All attributes are
preserved.
}
\examples{
cos_(1:5)
}
\seealso{
\code{\link[=cos]{cos()}}, \code{\link[=sin]{sin()}}, \code{\link[=tan]{tan()}}, \code{\link[=acos]{acos()}}, \code{\link[=asin]{asin()}}, \code{\link[=atan]{atan()}}, \code{\link[=atan2]{atan2()}}
}
