% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Trig_}
\alias{Trig_}
\alias{cos_}
\alias{sin_}
\alias{tan_}
\alias{acos_}
\alias{asin_}
\alias{atan_}
\alias{cospi_}
\alias{sinpi_}
\alias{tanpi_}
\title{Trigonometric Functions (Fast Parallel Version)}
\usage{
cos_(x, para = 50000L)

sin_(x, para = 50000L)

tan_(x, para = 50000L)

acos_(x, para = 50000L)

asin_(x, para = 50000L)

atan_(x, para = 50000L)

cospi_(x, para = 50000L)

sinpi_(x, para = 50000L)

tanpi_(x, para = 50000L)
}
\arguments{
\item{x}{vector of numeric values}

\item{para}{the minimum length of \code{x} to use parallel computation (50000
by default)}
}
\value{
A numeric vector, matrix, or data frame with the transformed values.
}
\description{
Fast version of trigonometric functions (when vector size >= 50000). They
respectively compute the cosine, sine, tangent, arc-cosine, arc-sine,
arc-tangent, and the two-argument arc-tangent.

\code{cospi_(x)}, \code{sinpi_(x)}, and \code{tanpi_(x)}, compute \code{cos(pi*x)},
\code{sin(pi*x)}, and \code{tan(pi*x)}.
}
\details{
They are \strong{not} generic functions and do not process factor, Date, POSIXt,
difftime, complex, or S4 objects (use base R equivalent function instead).
Data frames are processed column-wise, providing each column is compatible.
All attributes are preserved.
}
\examples{
cos_(1:5)
}
\seealso{
\code{\link[=cos]{cos()}}, \code{\link[=sin]{sin()}}, \code{\link[=tan]{tan()}}, \code{\link[=acos]{acos()}}, \code{\link[=asin]{asin()}}, \code{\link[=atan]{atan()}}, \code{\link[=atan2]{atan2()}}
}
